+++ src/main/java/com/github/javafaker/Resolver.java
@@ -1,0 +1,5 @@
+package com.github.javafaker;
+
+public interface Resolver {
+    String resolve(String key);
+}