+++ src/main/java/com/github/javafaker/service/FakeValuesService.java
@@ -1,38 +1,45 @@
 package com.github.javafaker.service;
 
-import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang.WordUtils;
+import com.github.javafaker.Resolver;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.InputStream;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
-public class FakeValuesService {
+public class FakeValuesService implements FakeValuesServiceInterface {
     private static final char[] METHOD_NAME_DELIMITERS = {'_'};
     private final Map<String, Object> fakeValuesMap;
     private final RandomService randomService;
 
     @SuppressWarnings({"unchecked", "rawtypes"})
     public FakeValuesService(Locale locale, RandomService randomService) {
-        String languageCode = locale.getLanguage();
-        final InputStream stream = findStream("/" + languageCode + ".yml");
+        String filename = locale.getLanguage();
+        InputStream stream = findStream(filename);
+        if (stream == null) {
+            filename = filename + "-" + locale.getCountry();
+            stream = findStream(filename);
+        }
         if (stream == null) {
             throw new LocaleDoesNotExistException(String.format("%s could not be found, does not have a corresponding yaml file", locale));
         }
+
         Map valuesMap = (Map) new Yaml().load(stream);
-        valuesMap = (Map) valuesMap.get(languageCode);
+        valuesMap = (Map) valuesMap.get(filename);
         fakeValuesMap = (Map<String, Object>) valuesMap.get("faker");
         this.randomService = randomService;
     }
 
     private InputStream findStream(String filename) {
-        InputStream streamOnClass = getClass().getResourceAsStream(filename);
+        String filenameWithExtension =  "/" + filename + ".yml";
+        InputStream streamOnClass = getClass().getResourceAsStream(filenameWithExtension);
         if (streamOnClass != null) {
             return streamOnClass;
         }
-        return getClass().getClassLoader().getResourceAsStream(filename);
+        return getClass().getClassLoader().getResourceAsStream(filenameWithExtension);
     }
 
     /**
@@ -57,6 +64,30 @@ public class FakeValuesService {
     }
 
     /**
+     * Safely fetches a key.
+     *
+     * If the value is null, it will return an empty string.
+     *
+     * If it is a list, it will assume it is a list of strings and select a random value from it.
+     *
+     * Otherwise it will just return the value as a string.
+     *
+     * @param key
+     * @return
+     */
+    @SuppressWarnings({"unchecked", "rawtypes"})
+    public String safeFetch(String key) {
+        Object o = fetchObject(key);
+        if (o == null) return "";
+        if (o instanceof List) {
+            List<String> values = (List<String>) o;
+            return values.get(randomService.nextInt(values.size()));
+        } else {
+            return (String) o;
+        }
+    }
+
+    /**
      * Return the object selected by the key from yaml file.
      *
      * @param key key contains path to an object. Path segment is separated by
@@ -71,42 +102,6 @@ public class FakeValuesService {
             currentValue = ((Map<String, Object>) currentValue).get(pathSection);
         }
         return currentValue;
-    }
-
-    /**
-     * A property that is composed of other properties.
-     * <p/>
-     * It firstly fetches the formatKey using {@link #fetch(String)}. It will
-     * proceed to convert the returned properties from the {@link #fetch(String)}
-     * method to a methodName and invoke this method against the object passed in.
-     * Finally, concatenation occurs with the return values of the methods
-     * using the joiner parameter as a separator.
-     *
-     * @param formatKey
-     * @param joiner
-     * @param objectToInvokeMethodOn
-     * @return
-     */
-    @SuppressWarnings({"unchecked", "rawtypes"})
-    public String composite(String formatKey, String joiner, Object objectToInvokeMethodOn) {
-        List<String> format = (List<String>) fetch(formatKey);
-
-        String[] parts = new String[format.size()];
-        for (int i = 0; i < parts.length; i++) {
-            // remove leading colon
-            String methodName = format.get(i).substring(1);
-            // convert to camel case
-            methodName = WordUtils.capitalizeFully(methodName, METHOD_NAME_DELIMITERS).replaceAll("_", "");
-            methodName = methodName.substring(0, 1).toLowerCase() + methodName.substring(1);
-
-            try {
-                parts[i] = (String) objectToInvokeMethodOn.getClass().getMethod(methodName, (Class[]) null).invoke(objectToInvokeMethodOn);
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        return StringUtils.join(parts, joiner);
     }
 
     /**
@@ -167,4 +162,32 @@ public class FakeValuesService {
     private int nextInt(int n) {
         return randomService.nextInt(n);
     }
+
+    /**
+     * Resolves a key to a method on an object.
+     *
+     * #{hello} with result in a method call to current.hello();
+     *
+     * #{Person.hello_someone} will result in a method call to person.helloSomeone();
+     *
+     * @param key
+     * @param current
+     * @param resolver
+     * @return
+     */
+    public String resolve(String key, Object current, Resolver resolver) {
+        String unresolvedString = safeFetch(key);
+        String regex = "#\\{[A-Za-z_.]+\\}";
+        Matcher matcher = Pattern.compile(regex).matcher(unresolvedString);
+        while (matcher.find()) {
+            String matched = matcher.group();
+            String strippedMatched = matched.replace('#', ' ').replace('{', ' ').replace('}', ' ').trim();
+            boolean isFirstLetterCapital = Character.isUpperCase(strippedMatched.charAt(0));
+            String objectWithMethodToResolve = isFirstLetterCapital ? strippedMatched : current.getClass().getSimpleName() + "." + strippedMatched;
+            String resolvedValue = resolver.resolve(objectWithMethodToResolve);
+            unresolvedString = unresolvedString.replace(matched, resolvedValue);
+        }
+        return unresolvedString;
+    }
+
 }