+++ src/main/java/com/github/javafaker/service/FakeValuesServiceInterface.java
@@ -1,0 +1,22 @@
+package com.github.javafaker.service;
+
+import com.github.javafaker.Resolver;
+
+public interface FakeValuesServiceInterface {
+
+    Object fetch(String key);
+
+    String fetchString(String key);
+
+    String safeFetch(String key);
+
+    Object fetchObject(String key);
+
+    String numerify(String numberString);
+
+    String bothify(String string);
+
+    String letterify(String letterString);
+
+    String resolve(String key, Object current, Resolver resolver);
+}