+++ src/test/java/com/github/javafaker/FakerTest.java
@@ -3,7 +3,12 @@ package com.github.javafaker;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.util.Locale;
+
 import static com.github.javafaker.matchers.MatchesRegularExpression.matchesRegularExpression;
+import static org.hamcrest.Matchers.isEmptyString;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.core.IsNot.not;
 import static org.junit.Assert.assertThat;
 
 public class FakerTest {
@@ -38,5 +43,15 @@ public class FakerTest {
     @Test
     public void numerifyShouldLeaveNonSpecialCharactersAlone() {
         assertThat(faker.numerify("####123"), matchesRegularExpression("\\d{4}123"));
+    }
+
+    @Test
+    public void resolveShouldReturnValueThatExists() {
+        assertThat(faker.resolve("address.city_prefix"), not(isEmptyString()));
+    }
+
+    @Test(expected = RuntimeException.class)
+    public void resolveShouldThrowExceptionWhenPropertyDoesntExist() {
+        faker.resolve("address.nothing");
     }
 }